requirements.txt
Package             Version
------------------- -------
Django              2.1.1  
djangorestframework 3.8.2  
Pillow              5.2.0  
pip                 18.0   
pytz                2018.5 
setuptools          40.4.2 
wheel               0.31.1 